SMD v8 Update! - Hydra updates   - Added the '@' '@:/' and '@/' path alias for the LOCAL VM directory ("./", the VM directory)   - Added the 'HYDRA:/' and ':/' path alias for the user's Hydra directory (one above the directory, can be used for resources)   - Fixed path normalazation to remove empty path seperators and bad path slashes - Sysusers Audit   - Checks the config for systemd-sysusers for any unneeded user/groups that could be deleted.   - Runs by default under a pacman hook, post hook after the same criteria for systemd-sysusers to run. - Updated some syscheck.d scripts to match the current permission handling scheme/format - Merged the two blocklists for modprobe together - Added additional sandbox/security settings for most systemd units (where applicable) - Updated Firefox/Librewolf config - Updated compression details for mkinitcpio

+++ usr/lib/smd/lib/modules/background.py
@@ -43,7 +43,7 @@ from lib.util.exec import stop, nulexec
 from os import mkdir, unlink, listdir, symlink
 from lib.util import nes, num, boolean, cancel_nul
 from lib.util.file import expand, hash_file, remove_file
-from os.path import isdir, isfile, islink, lexists, splitext
+from os.path import isdir, exists, isfile, islink, lexists, splitext
 from lib.constants.config import (
     DIRECTORY_LIBEXEC,
     BACKGROUND_PATH_CACHE,
@@ -222,6 +222,10 @@ class Background(object):
                     f'[m/background]: Cannot unlink the symlink path "{self._lockscreen}"!',
                     err,
                 )
+        if exists(self._lockscreen) and not islink(self._lockscreen):
+            return server.error(
+                f'[m/background]: Not removing non-symlink file "{self._lockscreen}"!'
+            )
         remove_file(self._lockscreen, True)
         try:
             symlink(image, self._lockscreen)
